trying 'table_exists'

+++ priv/repo/migrations/20250123025659_create_feedback_upvotes.exs
@@ -2,15 +2,21 @@ defmodule Onestack.Repo.Migrations.CreateFeedbackUpvotes do
   use Ecto.Migration
 
   def change do
-    execute "DROP TABLE IF EXISTS feedback_upvotes"
-
     create table(:feedback_upvotes) do
-      add :user_id, references(:users, on_delete: :delete_all), null: false
-      add :feedback_id, references(:feedbacks, on_delete: :delete_all), null: false
+      unless table_exists?(:feedback_upvotes) do
+        add :user_id, references(:users, on_delete: :delete_all), null: false
+        add :feedback_id, references(:feedbacks, on_delete: :delete_all), null: false
 
-      timestamps()
+        timestamps()
+      end
+
+      create unique_index(:feedback_upvotes, [:user_id, :feedback_id])
     end
+  end
 
-    create unique_index(:feedback_upvotes, [:user_id, :feedback_id])
+  defp table_exists?(table_name) do
+    query = "SELECT count(*) FROM sqlite_master WHERE type='table' AND name='#{table_name}';"
+    %{rows: [[count]]} = repo().query!(query)
+    count > 0
   end
 end
